refactor: StationPairDao 클래스 생성에 따른 기존 코드 분리
