+++ src/main/java/subway/application/LineService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.dao.LineDao;
 import subway.dao.SectionDao;
+import subway.dao.StationPairDao;
 import subway.domain.*;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
@@ -18,10 +19,12 @@ public class LineService {
 
     private final LineDao lineDao;
     private final SectionDao sectionDao;
+    private final StationPairDao stationPairDao;
 
-    public LineService(final LineDao lineDao, final SectionDao sectionDao) {
+    public LineService(final LineDao lineDao, final SectionDao sectionDao, final StationPairDao stationPairDao) {
         this.lineDao = lineDao;
         this.sectionDao = sectionDao;
+        this.stationPairDao = stationPairDao;
     }
 
     @Transactional
@@ -59,7 +62,7 @@ public class LineService {
 
     public LineWithStationsResponse findLineResponseById(Long id) {
         final Line persistLine = findLineById(id);
-        final List<StationPair> stationPairs = lineDao.findAllStationPair(id);
+        final List<StationPair> stationPairs = stationPairDao.findAllStationPair(id);
         final Stations stations = new Stations(stationPairs);
         return LineWithStationsResponse.of(persistLine, stations.getStations());
     }
