+++ src/main/java/subway/dao/LineDao.java
@@ -5,8 +5,6 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import subway.domain.Line;
-import subway.domain.StationPair;
-import subway.domain.Station;
 
 import javax.sql.DataSource;
 import java.util.HashMap;
@@ -16,26 +14,15 @@ import java.util.Optional;
 
 @Repository
 public class LineDao {
+
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert insertAction;
 
-    private RowMapper<Line> lineRowMapper = (rs, rowNum) ->
+    private final RowMapper<Line> lineRowMapper = (rs, rowNum) ->
             new Line(
                     rs.getLong("id"),
                     rs.getString("name"),
                     rs.getString("color")
-            );
-
-    private RowMapper<StationPair> stationPairRowMapper = (rs, rowNum) ->
-            new StationPair(
-                    new Station(
-                            rs.getLong("s1_id"),
-                            rs.getString("s1_name")
-                    ),
-                    new Station(
-                            rs.getLong("s2_id"),
-                            rs.getString("s2_name")
-                    )
             );
 
     public LineDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
@@ -81,14 +68,4 @@ public class LineDao {
                 .findAny();
     }
 
-    public List<StationPair> findAllStationPair(final Long lineId) {
-        String sql = "select s1.id as s1_id, s1.name as s1_name, s2.id as s2_id, s2.name as s2_name " +
-                "from section " +
-                "join station as s1 " +
-                "on section.up_station_id = s1.id " +
-                "join station as s2 " +
-                "on section.down_station_id = s2.id " +
-                "where line_id = ?";
-        return jdbcTemplate.query(sql, stationPairRowMapper, lineId);
-    }
 }
